Move code to new `.actor_runtime` and `.io`
