+++ pylib-v1/ailets/cons/environment.py
@@ -1,11 +1,11 @@
 from typing import Any, Dict
-from ailets.cons.atyping import IEnvironment, INodeRegistry
+from ailets.atyping import IEnvironment, INodeRegistry
 from ailets.cons.dagops import Dagops
 from ailets.cons.notification_queue import NotificationQueue
 from ailets.cons.processes import Processes
 from ailets.cons.seqno import Seqno
-from ailets.cons.piper import Piper
-from ailets.cons.memkv import MemKV
+from ailets.io.piper import Piper
+from ailets.io.memkv import MemKV
 
 
 class Environment(IEnvironment):
