+++ pylib-v1/ailets/models/gpt4o/__init__.py
@@ -1,4 +1,4 @@
-from ailets.cons.atyping import NodeDesc, Dependency
+from ailets.atyping import NodeDesc, Dependency
 
 messages_to_query = NodeDesc(
     name="messages_to_query",
