+++ pylib-v1/ailets/stdlib/query.py
@@ -2,9 +2,9 @@ import json
 import aiohttp
 import os
 import re
-from ailets.cons.atyping import INodeRuntime, StdHandles
+from ailets.atyping import INodeRuntime, StdHandles
 from ailets.cons.util import write_all
-from ailets.cons.input_reader import read_all
+from ailets.io.input_reader import read_all
 
 
 MAX_RUNS = 3  # Maximum number of runs allowed
